+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+### Fixed
+- PNDA-3016: When kafka topics have been deleted stop showing them in the PNDA console.
+
 ## [0.2.0] 2017-06-29
 ### Added:
 - PNDA-2691: Refactor for offline installation